+++ src/components/StreamGraphPanel.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { LoadingState } from '@grafana/data';
 import { StreamGraphPanel } from './StreamGraphPanel';
-import { ColorScheme, CurveType, LegendPlacement, StackOffset, StackOrder } from '../types';
+import { ColorScheme, CurveType, StackOffset, StackOrder } from "../types";
 import { multiDataFrames } from '../__mocks__/models/multiDataFrames';
 
 const baseOptions = {
@@ -13,8 +13,7 @@ const baseOptions = {
   fillOpacity: 0.8,
   showXAxis: true,
   showTooltip: true,
-  showLegend: true,
-  legendPlacement: LegendPlacement.BOTTOM,
+  legend: { showLegend: true, placement: 'bottom' as const, displayMode: 'list' as const },
 };
 
 const baseTimeRange = {